General edit
The tag descriptions were changed to make them more meaningful
(<em></em> around the description of the italics tag, for instance), and
the new header tag was incorporated into the tag list and the examples.
An explanation for the updated usage was added, and the whole readme was
generally edited for clarity and structure.